Promote StylePropertySpecification
It's useful outside of expressions.
